Move known samplers to HlslKnownMethods type
